# docs/swagger.yaml
definitions:
...
  http.UpdateProductReq:
    properties:
      category_ids:
        items:
          type: string
        type: array
      id:
        type: string
      media_ids:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      reorder_level:
        type: integer
      reorder_quantity:
        type: integer
      stock_level:
        type: integer
    required:
    - category_ids
    - id
    - media_ids
    - name
    - price
    - reorder_level
    - reorder_quantity
    - stock_level
    type: object
...
  http.categoryObject:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  http.createProductReq:
    properties:
      category_ids:
        items:
          type: string
        type: array
      media_ids:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      reorder_level:
        type: integer
      reorder_quantity:
        type: integer
      stock_level:
        type: integer
    required:
    - category_ids
    - name
    - price
    - reorder_level
    - reorder_quantity
    - stock_level
    type: object
...
  http.deleteProductRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
  http.detailProductResp:
    properties:
      avatar:
        items:
          $ref: '#/definitions/internal_shop_delivery_http.avatar_obj'
        type: array
      category_id:
        items:
          type: string
        type: array
      category_name:
        items:
          type: string
        type: array
      id:
        type: string
      inventory_name:
        type: string
      name:
        type: string
      price:
        type: number
      shop_id:
        type: string
      shop_name:
        type: string
    required:
    - category_id
    - category_name
    - id
    - inventory_name
    - name
    - price
    - shop_id
    - shop_name
    type: object
...
  http.getProductResp:
    properties:
      products:
        items:
          $ref: '#/definitions/http.listProductItem'
        type: array
      shop_object:
        $ref: '#/definitions/internal_shop_delivery_http.shopObject'
    type: object
...
  http.inventoryObject:
    properties:
      id:
        type: string
      name:
        type: string
      stock_level:
        type: integer
    type: object
...
  http.listProductItem:
    properties:
      avatar:
        items:
          $ref: '#/definitions/internal_shop_delivery_http.avatar_obj'
        type: array
      category_objects:
        items:
          $ref: '#/definitions/http.categoryObject'
        type: array
      id:
        type: string
      inventory_object:
        $ref: '#/definitions/http.inventoryObject'
      name:
        type: string
      price:
        type: number
      shop_id:
        type: string
    type: object
...
  http.updateProductResp:
    properties:
      avatar:
        items:
          $ref: '#/definitions/internal_shop_delivery_http.avatar_obj'
        type: array
      category_id:
        items:
          type: string
        type: array
      category_name:
        items:
          type: string
        type: array
      id:
        type: string
      inventory_name:
        type: string
      name:
        type: string
      price:
        type: number
      shop_id:
        type: string
      shop_name:
        type: string
    required:
    - id
    type: object
...
  internal_shop_delivery_http.avatar_obj:
    properties:
      media_id:
        type: string
      url:
        type: string
    type: object
...
  internal_shop_delivery_http.shopObject:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  response.Resp:
    properties:
      data: {}
      error_code:
        type: integer
      errors: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/shops/create-product:
    post:
      consumes:
      - application/json
      description: Create a new product in the shop
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/response.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Create a product
      tags:
      - Shop
...
  /api/v1/shops/products/{id}:
    get:
      consumes:
      - application/json
      description: Get shop detail by id
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/http.detailProductResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: get shop detail by their id
      tags:
      - Products
...
  /api/v1/shops/products/delete:
    delete:
      consumes:
      - application/json
      description: delete product by id
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.deleteProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: delete product by their id
      tags:
      - Products
  /api/v1/shops/products/get-product:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of products with optional filters , at
        least 1 parameter request body must be provided
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: category_ids
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: ids
        type: array
      - in: query
        name: search
        type: string
      - in: query
        name: shop_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated Products
          schema:
            $ref: '#/definitions/http.getProductResp'
      summary: Get products with pagination
      tags:
      - Product
  /api/v1/shops/products/update:
    post:
      consumes:
      - application/json
      description: Update product BY ID , only ID is required , other fields are optional
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/http.updateProductResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Update product
      tags:
      - Products
...
swagger: "2.0"